requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
langgraph
pandas
beautifulsoup4
html2text
python-dotenv